requirements.txt
frida==12.8.0
frida-tools==5.3.0
objection==1.8.4
Appium-Python-Client
lxml